requirements.txt
pandas==1.1.2
numpy==1.19.2
requests==2.24.0
scikit-learn==0.23.2
python-dotenv==0.14.0

pylint==2.6.0
mypy==0.782
flake8==3.8.3